chore(lepton): reorganize header includes and implement core class and object structures
