+++ shared/include/lepton/core.h
@@ -3,8 +3,8 @@
 #include <size.h>
 #include <int.h>
 
-#include <lepton/class.h>
-#include <lepton/object.h>
+#include <lepton/core/class.h>
+#include <lepton/core/object.h>
 #include <lepton/result.h>
 #include <lepton/throwable.h>
