+++ shared/include/lepton/core/object.h
